requirements.txt
text
fastapi
uvicorn[standard]
python-dotenv
requests
python-levenshtein
pydantic
overrides
# optional libpostal binding
postal